util: Add read_file() function for reading sysfs files
Add a function that can read a complete, unknown size file for
reading entry point files from sysfs.  This function models its signature
on the mem_chunk() funtion, so it also allocates memory that the caller
needs to free.  The files that we are interested in reading are very small,
and have a known upper bound on the size.  The EINTR handling is based
on the myread() function.

Contributed by Roy Franz.