dmidecode: Add passing of flags parameter to dmi_table()
This allows dmi_table() to print the address that the table was read from,
even when reading from a sysfs file. Previous to sysfs support, the address
of the table was always the same as the offset within the file it was read
from (usually /dev/mem.)  For sysfs files that contain just the table at
offset 0, we still want to pass the address and display that as normal,
but read from offset 0 in the file. If the flag FLAG_NO_FILE_OFFSET
is passed, dmi_table() uses 'base' as the DMI table address for display,
but uses 0 as the file offset.

Contributed by Roy Franz.