dmidecode: Add reading of SMBIOS tables from sysfs
Add preferential reading of the SMBIOS tables from
/sys/firmware/dmi/tables.  If these files are not present or
not valid, the previously supported methods of locating SMBIOS
tables are attempted.

Messages indicating which source is used for the tables
have been added.  These are printed before the tables have
been validated so they can go at the top of the output.  This
also shows what methods have been tried and failed due to invalid
tables.

The address of the entry point is not known when read from sysfs,
so it is not printed in that case.

A placeholder print is added where 64-bit entry point processing
will be added.

Contributed by Roy Franz.