dmidecode: Display types 41 and 42 in quiet mode too
Originally the highest regular DMI structure type was 39, and non-OEM
types greater than that (including 126/disabled and 127/end-of-table)
wouldn't show up in quiet mode. The limit was not updated when types
41 and 42 were added to the SMBIOS specification, so these types do
not show up in quiet mode. Fix that.

We could simply raise the limit from 39 to 42, but the fact that this
bug went unnoticed for years is probably a good indication that this
limit was a bad idea in the first place and it is safer to only
explicitly hide types 126 and 127 in quiet mode.