dmidecode: Get OEM vendor from System Information (DMI type 1)
DMI type 0 is the BIOS information, it doesn't tell us about the
system vendor. DMI type 1 does that. It made no difference so far
because the only vendor with support for OEM-type decoding was HP and
they make their own BIOS, but it will matter as soon as we add support
for more vendors.