dmidecode: Use read_file() to read the DMI table from sysfs
We shouldn't use mem_chunk() to read the DMI table from sysfs. This
will fail for SMBIOS v3 implementations which specify a maximum length
for the table rather than its exact length. The kernel will trim the
table to the actual length, so the DMI file will be shorter than the
length announced in entry point.

read_file() fits the bill in this case, as it deals with end of file
nicely.

This also helps with corrupted DMI tables, as the kernel will not
export the part of the table that it wasn't able to parse, effectively
trimming it.

This fixes bug #46176:
https://savannah.nongnu.org/bugs/?46176
Unexpected end of file error