Goodbye CHANGELOG, welcome NEWS
The CHANGELOG file as it exists today is essentially redundant with
git log. It's not really user friendly, and duplicates the work on
the committer side.

Delete that file, and replace it with a more structured NEWS file,
listing only the most relevant changes in each version. The contents
would be pretty much the same as the release announcements.

This initial version covers most past releases.