dmidecode: Don't allocate more memory than needed
If the actual DMI table size is less than the announced maximum
(which is allowed for 64-bit SMBIOS 3 entry points), we may allocate
significantly more memory than is actually needed. If reading from
/dev/mem, there's nothing we can do about that. However, is reading
from sysfs or from a dump file, we can easily check the file size
and compute the actual table size from it. That way we only allocate
the required amount of memory.

Credits to Lionel Debroux for seeding the idea when performing fuzz
testing on dmidecode.

Signed-off-by: Jean Delvare <jdelvare@suse.de>