util: Don't close the same file descriptor twice
If myread() fails, it closes the file descriptor it was reading from.
On success, it does not. The caller is always closing the file
descriptor afterward, which results in double close in the error
case. This causes an additional error:

/dev/mem: Bad file descriptor

As the caller is always taking care of closing the file descriptor,
there is no need for myread() to do it.

Signed-off-by: Jean Delvare <jdelvare@suse.de>
Fixes: 458f73d58c24 ("Fix error paths in mem_chunk")