dmidecode: SMBIOS data is little-endian
I originally assumed that the SMBIOS data would use the byte ordering
of the host, however since v2.8.0 the SMBIOS specification clarified
that the SMBIOS data is always little-endian. So change the access
helpers accordingly.

This is good news as far as portability is concerned: we can dump the
table of one host and decode it on another.

Signed-off-by: Jean Delvare <jdelvare@suse.de>