dmidecode: Sanity check the table offset in dump files
If the offset (base) is beyond the end of the file (statbuf.st_size),
the computations will lead to an integer overflow. As it doesn't make
sense in the first place, check for this condition and fail
immediately.

This bug was discovered by Lionel Debroux using the AFL fuzzer and
AddressSanitizer.

Signed-off-by: Jean Delvare <jdelvare@suse.de>
Fixes: bd78a5dfd470 ("dmidecode: Don't allocate more memory than needed")