dmidecode: Adding bios-revision to -s option
Some hardware vendors like HPe use the Version field to store the
bios generation like (U30, U32, A40, ...).  If you want to get the
"release" version of this bios generation, the bios revision field
must be considered.

A typical output of this kind of server looks like :

          BIOS Information
                  Vendor: HPE
                  Version: A40
                  Release Date: 07/20/2019
                  [...]
                  BIOS Revision: 2.0
                  Firmware Revision: 1.45

Add a "bios-revision" entry into the '-s' option.

A typical usage of this feature looks like :

	[root@host] dmidecode -s bios-revision
	2.0

Signed-off-by: Erwan Velu <e.velu@criteo.com>
Signed-off-by: Jean Delvare <jdelvare@suse.de>