dmioem: Fix vendor string comparison
The vendor string comparison code made an incorrect assumption about
the length of the vendor string found in the DMI table. We would hit
a false positive if that string was shorter than, and an exact prefix
of, the vendor string we were comparing with.

Explicitly check for length equality before comparing the strings.
Also rewrite the whole thing as an iteration over a table, instead
of open-coding it. This will make it easier to add support for more
vendors in the future.

Signed-off-by: Jean Delvare <jdelvare@suse.de>