dmioem: HPE type 242: Fix ID on 32-bit systems
I noticed that the reported ID is different when building dmidecode
in 32-bit mode. The reason is that %lx isn't a valid format to print
a 64-bit integer. It works by accident on 64-bit systems because
types "long integer" and "long long integer" are the same there, but
on 32-bit systems, only the latter can actually store a 64-bit value.

Change the format to %llx so that it works on 32-bit systems too.

Signed-off-by: Jean Delvare <jdelvare@suse.de>
Fixes: 90d1323a8d14 ("dmioem: Decode HPE OEM Record 242")
Cc: Jerry Hoemann <jerry.hoemann@hpe.com>